--- carousel.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-  <meta charset="UTF-8"/>
-  <title>carousel</title>
-  <script src="./dist/css.js"></script>
-  <link href="//maxcdn.bootstrapcdn.com/font-awesome/4.1.0/css/font-awesome.min.css" rel="stylesheet" />
-</head>
-<body>
-  <div id="wrap"><div id="slide"></div></div>
-  <script src="./dist/js.js"></script>
-</body>
-</html>
