--- carousel/README.md
@@ -1,2 +1,0 @@
-# javascript-amazon
-레벨3
