--- carousel/server/src/carouselData.json
@@ -1,36 +1,0 @@
-{
-  "navData": [
-    "Fast, Free Delivery",
-    "Exclusive deals and offers",
-    "Prime Originals, movies and TV shows",
-    "Over 2 million song ad free"
-  ],
-  "contentData": [
-    {
-      "imgUrl": "https://imgur.com/eRJL1hm.jpg",
-      "title": "Enjoy exclusive savings and rewards",
-      "content": [
-        "1Save on groceries and get an additional 10% off select sale items in all US Whole Food Market stores",
-        "Save 20% off diapers, baby food, and more",
-        "Earn 5% back at AMazon.com with the Amazon Prime Rewards Visa Card",
-        "Get exclusive deals and discounts and 30-minute early access to Lightning Deals"
-      ]
-    },
-    {
-      "imgUrl": "https://imgur.com/LLxEXCq.jpg",
-      "title": "title2",
-      "content": ["content1", "content2"]
-    },
-    {
-      "imgUrl": "https://imgur.com/tq3t2xx.jpg",
-      "title": "title3",
-      "content": ["content1", "content2", "content3"]
-    },
-    {
-      "imgUrl": "https://imgur.com/eRJL1hm.jpg",
-      "title": "title4",
-      "content": ["content1", "content2", "content3", "content4"]
-    }
-  ],
-  "buttonData": ["prev", "next"]
-}
