--- carousel/src/js/main.js
@@ -1,5 +1,0 @@
-import Controller from '../../controller/controller.js'
-
-window.addEventListener("DOMContentLoaded", () => {
-  new Controller()
-})
