--- carousel/utlis.js
@@ -1,5 +1,0 @@
-export const $ = el => document.querySelector(el)
-export const $$ = el => document.querySelectorAll(el)
-
-
-// 전역변수를 없애려고 객체를 활용 할 예정입니다 const hueyUtils = { $() {}, .....}
