+++ client/src/view/searchBarView.js
