--- package.json
@@ -1,29 +1,0 @@
-{
-  "name": "javascript-amazon",
-  "version": "1.0.0",
-  "description": "레벨3",
-  "main": "main.js",
-  "scripts": {
-    "build": "webpack --config webpack.config.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Hu2y/javascript-amazon.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@babel/cli": "^7.8.4",
-    "axios": "^0.19.2"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.2.2",
-    "@babel/preset-env": "^7.3.1",
-    "babel-loader": "^8.0.5",
-    "css-loader": "^3.4.2",
-    "style-loader": "^1.1.3",
-    "webpack": "^4.41.6",
-    "webpack-cli": "^3.2.1",
-    "webpack-dev-server": "^3.1.14"
-  }
-}
