--- search/server/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "amazon_test",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "node -r esm src"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@koa/cors": "^3.0.0",
-    "axios": "^0.19.2",
-    "esm": "^3.2.25",
-    "firebase-admin": "^8.9.2",
-    "firebase-functions": "^3.3.0",
-    "koa": "^2.11.0",
-    "koa-firebase-functions": "^1.0.1",
-    "koa-router": "^8.0.8"
-  }
-}
