--- search/server/src/index.js
@@ -1,4 +1,0 @@
-/* eslint-disable no-global-assign */
-
-require = require('esm')(module /*, options */);
-module.exports = require('./main.js');
