+++ server/src/index.js
@@ -1,4 +1,4 @@
 /* eslint-disable no-global-assign */
 
-require = require('esm')(module /*, options */)
-module.exports = require('./main.js')
+require = require('esm')(module /*, options */);
+module.exports = require('./main.js');
