+++ server/src/main.js
@@ -1,20 +1,31 @@
-import Koa from 'koa'
-import Router from 'koa-router'
+const Koa = require('koa')
+const Router = require('koa-router')
+const functions = require('firebase-functions')
 const cors = require('@koa/cors');
-const localData = require('./carouselData.json')
+const searchData = require('./searchData.json')
 
-const app = new Koa()
-const router = new Router()
+const app = new Koa();
+const router = new Router();
 
+console.log('a')
 app.use(cors());
 
 router.get('/', (ctx, next) => {
-  ctx.body = localData
+  ctx.body = searchData;
 });
 
-app.use(router.routes())
+app.use(router.routes());
 
-const port = 4000
-app.listen(port, () => {
-  console.log('listen to port %d', port)
-})
+const port = 4000;
+
+// app.listen(port, () => {
+//   console.log('listen to port %d', port);
+// })
+
+const server = app.listen(port, () => {
+  console.log(`HITMers-server is running on port ${port}`);
+});
+
+module.exports = server;
+exports.api = functions.https.onRequest(app.callback());
+
