--- src/style/main.js
@@ -1,2 +1,0 @@
-const resetCss = require('./reset.css')
-const styleCss = require('./style.css')
